update -lettre - notifacation - autre f°

+++ src/Controller/UserController.php
@@ -20,16 +20,24 @@ use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 class UserController extends AbstractController
 {
     #[Route('/', name: 'app_user_index', methods: ['GET'])]
-    public function index(UserRepository $userRepository): Response
+    public function index(UserRepository $userRepository ,Request $request): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN", null,"");
+
+
+       $search=  $request->get('search');
+       $searched =   $userRepository->search($search);
+
+            if($search) {
+                 $users = $userRepository->search($search);
+            } else {
+                $users = $userRepository->findAll();
+            }
+
+
+
         return $this->render('user/index.html.twig', [
-            'users' => $userRepository->findAll(),
-            'imageField' => ImageField::new('image')
-            ->setLabel('Image')
-            ->setBasePath('images/user') // Chemin de base pour afficher les images
-            ->setUploadDir('public/images/user') // Répertoire de téléchargement des images
-            ->setUploadedFileNamePattern('[randomhash].[extension]') // Modèle de nom de fichier téléchargé
-            ->setRequired(false)
+            'users' => $users,
             
         ]);
     }
@@ -37,6 +45,8 @@ class UserController extends AbstractController
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request,UserPasswordHasherInterface $userPasswordHasher ,EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN", null,"");
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -66,20 +76,42 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
-        return $this->render('user/show.html.twig', [
-            'user' => $user,
-        ]);
+
+        if ($this->isGranted("ROLE_ADMIN")) {
+            return $this->render('user/show.html.twig', [
+                'user' => $user,
+            ]);
+    
+        }else{
+            return $this->render('user/show_client.html.twig', [
+                'user' => $user,
+            ]);
+    
+      
+         }
+
+
+       
+
+
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, UserPasswordHasherInterface $userPasswordHasher, User $user, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(UserType::class, $user);
+        
+         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
 
-           
+            $user->setPassword(
+                $userPasswordHasher->hashPassword(
+                    $user,
+                    $form->get('password')->getData()
+                )
+            );
+
            
             $entityManager->persist($user);
             $entityManager->flush();
@@ -87,16 +119,27 @@ class UserController extends AbstractController
 
             return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
         }
+        if ($this->isGranted("ROLE_ADMIN")) {
+            return $this->render('user/edit.html.twig', [
+                'user' => $user,
+                'form' => $form->createView(),
+            ]);
+        }else{
+            return $this->render('user/edit_client.html.twig', [
+                'user' => $user,
+                'form' => $form->createView(),
+            ]);
+      
+        }  
 
-        return $this->render('user/edit.html.twig', [
-            'user' => $user,
-            'form' => $form->createView(),
-        ]);
+       
     }
 
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN", null,"");
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
